ui: shorten week header and streak text for mobile
- 'Week of January 19th' → 'Week of Jan 19th'
- '2 day streak' → '2 days' (with singular/plural handling)

+++ src/utils/habitSchedule.js
@@ -286,7 +286,7 @@ export const getFullDayName = (date) => {
  */
 export const formatWeekOfHeader = (weekStart) => {
   const day = weekStart.getDate()
-  const month = weekStart.toLocaleDateString('en-US', { month: 'long' })
+  const month = weekStart.toLocaleDateString('en-US', { month: 'short' })
 
   // Add ordinal suffix
   const suffix = (day === 1 || day === 21 || day === 31) ? 'st'
